extract provider fs-layout to providerdir
