+++ dhctl/pkg/infrastructureprovider/providerdir/providerdir.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package providerdata
+package providerdir
 
 import (
 	"path/filepath"
